Adds a parameter to set the start index of lists
This commit changes the `List` class to accept one additional parameter
`start_index` with default value `1`, of type `int`. The parameter is
used to shift the index that is sent to `_get_bullet_layout_element` and
is used by that said method in the `OrderedList` class. This will allow
for lists to continue the numbering across paragraph sections or from
one table cell to the next.

Features
- Add `start_index` parameter to list objects.